Add tests and refactor
