+++ packages/core/data-transfer/lib/bootstrap/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from './push';
+export { default as createTransferController } from './transfer';
